Customize flow no longer expects non-standard changes to ringtone picker.
The RingtonePickerActivity was previously modified to send back the
ringtone title with the returned intent.  Since we reverted
RingtonePickerActivity to android-1.5r2, this feature is no longer
present and should not be used.

Instead, we use RingtoneManager.getRingtone(this, uri).getTitle(this)
which works the same but is slightly less efficient.

CR: Sergey Ten