Included package name in PackageResources URIs.
This is a first step toward supporting theme removal gracefully within
the ProfileManager.  URIs can be left dangling to themes that have been
replaced or removed and upon application or preview of the profile we
can detect that the Uri yields an empty cursor and therefore should be
treated as a null or default fallback value.

This code needs more testing and further exploration.

CR: Ed Carrigan