Introduced support for HTC's lock screen wallpapers.
This support is currently not functioning as it appears HTC's
implementation does not behave as initially documented.  Pending further
discussion and review from HTC.

Also adds a dependency on ProfileManager in order to determine the
currently applied scene id.  This can be detected by using HTC's
provider but this would introduce a potential race condition on profile
swap as we broadcast a request to change the value to both Rosie and
ThemeManager at the same time.

CR: Ed Carrigan