ESPRESSO-137: Improved perceived performance of theme change.
This change cooperates with another change in ProfileManager which is
designed to improve perceived performance of theme, style, and
profile change.  This is accomplished by "holding" the user on the chooser
screen until the ACTION_THEME_CHANGED broadcast completes (or a
configuration change event is received by that app).  This gives the
illusion of stability throughout the event as the user sees no changes
until it is complete.  At any time the user may leave this screen with
the Home key (or by another activity coming up) which will give them the
"raw" theme change experience.

Additionally it was found that not all receiver events (most notably
ChangeProfile and ChangeTheme) were not running in the background queue.
This caused starvation of the main thread which has now been fixed.

CR: Ed Carrigan