Fixed crash when [re]installing any third party APK.
This patch fixes the behaviour of ThemeManager when it improperly senses
a package removal or addition of a non-theme package.  This is not
normally supposed to happen, however, as it uses an intent filter which
only listens on the theme category from the PackageManager.  This
strongly suggests the PackageManager has a bug or regression which is
causing it to no longer discriminate theme packages from regular
packages.

CR: Dirk Sigurdson