ESPRESSO-930: Associate assets with default theme.
This causes a slight regression where we will duplicate the assets in
other chooser screens (ringtone picker and wallpaper chooser) because
some of the entries exist elsewhere.  This can be fixed longer term by
identifying stable URLs into existing providers (like via the
RingtoneManager) and using those instead of hardcoding new ones directly
to the file on /system.

Alternatively, we can modify the choosers to simply ignore the default
theme, but IMO, this is a lot clumsier.

CR: Dirk Sigurdson
CR: Ed Carrigan