Work around upstream race condition bug during provider onCreate.
There is a race condition in Google's code which causes threads started
by the provider's onCreate to not block waiting for the onCreate method
to finish.  Instead, a second ContentProvider instance will be created
on the thread which issued the request.

This bug has been reported to Google here:

http://code.google.com/p/android/issues/detail?id=7305

CR: Dirk Sigurdson
CR: Ed Carrigan